requirements.txt
flask==2.0.1
werkzeug==2.0.1
gunicorn==20.1.0
llama-cpp-python==0.2.11
requests==2.28.1
